Add tests for nginx version
I had considered doing something with regex like `\d+(\.\d+)*`, then I
realized that there are other valid version strings that perhaps do not
contain numerics separated by `.`, so this option is a bit more flexible
as far as what versions nginx can return.

The major weakness here is that it does require nginx to return the
version on the first line, and to use a `/` separator before the version
itself.

One potential alternative would be to use the regex
`r'(\w+\.(\w+\.?)*)'` and get groups()[0] from that, however that
*would* require at least a decimal point.

There's a number of other options that one could use here, however,
we're currently simply banking on nginx always returning the actual
version prefixed with a forward slash.