Add tests + ssl arg for mongodb_user state
These tests are not *super* meaningful but they do at least ensure that
ssl the arg correctly gets passed into the MongoClient. Still left to do
is add tests + ssl arg for the mongodb_database state, the mongodb
module, as well as updating all of the documentation for these modules.

We also want to add a few functional tests that ensure that the SSL
argument is correctly used with an actual mongodb server.