Fix flaky sdb/test_etcd_db test
It appears that the etcd service inside the container is not actually
running or something, despite the container running. Rather than simply
failing on the tests, inside the fixture let's try to connect to sdb. If
we cannot, then retry up to 10 times, and if that fails then we should
probably just skip the test since the failing test doesn't really reveal
anything salt related, just about our test infrastructure.