Skip `tests/unit/{modules,states}/test_zcbuildout.py` on windows.
It needs special work on the golden images to get SSL to properly work.
These steps are required because the code being tested is using
`easy_install` which does not know how to get certificates from `certifi`.
Since `easy_install` is too old, and deprecated, the extra work is not
worth it, plus, they are still being tested on other platforms.

Signed-off-by: Pedro Algarvio <palgarvio@vmware.com>